# Swagger.yaml
openapi: 3.0.3
info:
  title: PedeAi API
  description: API de gerenciamento de pedidos para restaurantes/lanchonetes.
  contact:
    name: Suporte PedeAi API
    url: http://pedeai-api.com/suporte
    email: contato@pedeai-api.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /itens:
    get:
      tags:
        - Item
      summary: Buscar itens
      description: Busca itens por categoria
      parameters:
        - name: categoria
          in: query
          description: Categoria do item
          schema:
            $ref: '#/components/schemas/Categoria'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Itens não encontrados
          content:
            application/json:
              schema:
                type: string
                example: Itens não encontrados
    post:
      tags:
        - Item
      summary: Criar item
      description: Cria um novo item
      requestBody:
        description: Dados do item para criação
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreateUpdateDto'
        required: true
      responses:
        '201':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UUID'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
                example: Bad Request
        '404':
          description: Item não encontrado
          content:
            application/json:
              schema:
                type: string
                example: Item não encontrado
        '409':
          description: Conflito
          content:
            application/json:
              schema:
                type: string
                example: Conflito
  /itens/{id}:
    get:
      tags:
        - Item
      summary: Buscar item por ID
      description: Busca um item pelo ID
      parameters:
        - name: id
          in: path
          description: ID do item
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
          example: 257ae14b-8bb7-4a80-9a68-22197f72ff47
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item não encontrado
          content:
            application/json:
              schema:
                type: string
                example: Item não encontrado
    put:
      tags:
        - Item
      summary: Atualizar item
      description: Atualiza um item existente
      parameters:
        - name: id
          in: path
          description: ID do item
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
          example: 63a59178-39f8-4a28-a2c7-989a57ca7b54
      requestBody:
        description: Dados do item para atualização
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreateUpdateDto'
        required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item não encontrado
          content:
            application/json:
              schema:
                type: string
                example: Item não encontrado
    delete:
      tags:
        - Item
      summary: Remover item
      description: Remove um item pelo ID
      parameters:
        - name: id
          in: path
          description: ID do item
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
          example: 082db643-11a4-4bf8-8115-72148e24262d
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: string
                example: Item removido com sucesso
        '404':
          description: Item não encontrado
          content:
            application/json:
              schema:
                type: string
                example: Item não encontrado
...
components:
  schemas:
    Categoria:
      enum:
        - LANCHE
        - ACOMPANHAMENTO
        - BEBIDA
        - SOBREMESA
      type: string
...
    Item:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        nome:
          type: string
        preco:
          format: int32
          type: integer
        categoria:
          $ref: '#/components/schemas/Categoria'
        descricao:
          type: string
        imagem:
          type: string
    ItemCreateUpdateDto:
      required:
        - nome
        - preco
        - categoria
        - descricao
        - imagem
      type: object
      properties:
        nome:
          type: string
          example: '1'
        preco:
          format: int32
          type: integer
          example: 1
        categoria:
          type: string
          allOf:
            - $ref: '#/components/schemas/Categoria'
          example: '1'
        descricao:
          type: string
          example: '1'
        imagem:
          type: string
          example: '1'
...
    UUID:
      format: uuid
      pattern: >-
        [a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}
      type: string